refactor: splitted Workflow mechanism
